fix ts errors
